requirements.txt
tensorflow
tensorflow-model-optimization
tensorflow-addons
tensorflow-recommenders
tensorflow-probability
onnxruntime
tf2onnx
torch
wandb
tqdm
numba
pydot
pydotplus
pydot_ng
graphviz
pdf2image
seaborn
overrides
transformers
lightgbm
xgboost
catboost
mmh3
deepctr
deepctr-torch
glove-py
fasttext
pandarallel
deepspeed
dask[complete]
pymp-pypi
opencv-python
scikit-image
ipywidgets
tfrecord-lite
qgrid
swifter
spektral
tf2_gnn
gensim
pyarrow
tabulate
omegaconf
icecream
jieba 
sentencepiece
kaggle
googletrans==4.0.0-rc1
spacy
datasets
yapf
wordcloud
iterative-stratification 
optuna
nni
wordcloud
cityhash
func-timeout
pytorch-fast-transformers
